requirements.txt
pdfplumber==0.11.4
sentence-transformers==3.1.1
faiss-cpu==1.8.0.post1
groq==0.11.0
python-dotenv==1.0.1
gradio==4.44.0